lambda_container/app/requirements.txt
numpy
opencv-python
pytesseract
pillow
PyMuPDF